Move SendTransactionServiceStats and Report to new module
