+++ send-transaction-service/src/lib.rs
@@ -1,6 +1,9 @@
 #![allow(clippy::arithmetic_side_effects)]
 pub mod send_transaction_service;
+pub mod send_transaction_service_stats;
 pub mod tpu_info;
+
+pub use send_transaction_service_stats::SendTransactionServiceStats;
 
 #[macro_use]
 extern crate solana_metrics;
